Create PushQueue class.

+++ src/main/kotlin/pathfinder/weatherBot/location/Climate.kt
@@ -16,10 +16,10 @@ enum class Climate(val winterTemp: Long, val springTemp: Long, val summerTemp: L
                 in 1..20 ->  ({-3 d 10} to (1 d 4))
                 in 21..40 ->  ({-2 d 10} to (1 d 6)+1)
                 in 41..60 ->  ({-1 d 10} to (1 d 6)+2)
-                in 61..80 ->  ({ 0.toLong() } to (1 d 6)+2)
+                in 61..80 ->  ({ 0L } to (1 d 6)+2)
                 in 81..95 ->  ({1 d 10} to (1 d 6)+1)
                 in 96..99 ->  ({2 d 10} to (1 d 4))
-                100.toLong() -> ({3 d 10} to (1 d 2))
+                100L -> ({3 d 10} to (1 d 2))
                 else -> throw dHundredException
             }
         }
@@ -36,7 +36,7 @@ enum class Climate(val winterTemp: Long, val springTemp: Long, val summerTemp: L
                 in 1..5 ->  ({-3 d 10} to (1 d 2))
                 in 6..15 ->  ({-2 d 10} to (1 d 4))
                 in 16..35 ->  ({-1 d 10} to (1 d 4)+1)
-                in 36..65 ->  ({0.toLong()} to (1 d 6)+1)
+                in 36..65 ->  ({0L} to (1 d 6)+1)
                 in 66..85 ->  ({1 d 10} to (1 d 4)+1)
                 in 86..95 ->  ({2 d 10} to (1 d 4))
                 in 96..100 -> ({3 d 10} to (1 d 2))
@@ -55,7 +55,7 @@ enum class Climate(val winterTemp: Long, val springTemp: Long, val summerTemp: L
             return when (1 d 100){
                 in 1..10 ->  ({ -2 d 10} to (1 d 2))
                 in 11..25 ->  ({-1 d 10} to (1 d 2))
-                in 26..55 ->  ({0.toLong()} to (1 d 4))
+                in 26..55 ->  ({0L} to (1 d 4))
                 in 56..85 ->  ({1 d 10} to (1 d 4))
                 in 86..100 -> ({2 d 10} to (1 d 2))
                 else -> throw dHundredException
